basic stem and pie charts added
